added playerControls file to house all control logic. Added gamecontroller js library to handle gamepad. Added initial gamepad logic
