+++ index.html
@@ -13,8 +13,7 @@
   </style>
 </head>
 <body>
-  <!-- <canvas id="game"></canvas> -->
-  <!-- <script src="https://kaboomjs.com/lib/0.1.0/kaboom.js"></script>\ -->
+  <script src="gamecontroller.min.js"></script>
   <script src="kaboom.js"></script>
   <script src="game.js" type="module"></script>
 </body>
